requirements.txt
amadeus==9.0.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
executing==2.0.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.3
traitlets==5.10.1
typing_extensions==4.7.1
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
